Типизировал redux
