+++ src/Components/Pagination/pagination.scss
@@ -1,0 +1,42 @@
+.pagination {
+  display: flex;
+
+  & li {
+    width: 40px;
+    height: 40px;
+    border: 2px solid transparent;
+    border-radius: 50%;
+    gap: 16px;
+
+    &:hover {
+      border-color: #ffdf8c;
+    }
+  }
+
+  & a {
+    display: flex;
+    width: 100%;
+    height: 100%;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+  }
+
+  & .selected {
+    border-color: #ffdf8c;
+  }
+
+  & .disabled {
+    opacity: 0.5;
+    cursor: none;
+
+    &:hover {
+      border-color: transparent;
+    }
+
+    & a {
+      cursor: default;
+    }
+  }
+}
