+++ src/Components/Sort/sort.tsx
@@ -4,12 +4,8 @@ import { useDispatch, useSelector } from "react-redux";
 import { filterSelector, setSort } from "../../Redux/filterSlice";
 
 import { sortList, SortListItemType } from "./sortList.enum";
-import './sort.scss';
 
-type PopupClickType = MouseEvent & {
-  path: Node[];
-  composedPath: [];
-};
+import './sort.scss';
 
 const Sort: React.FC = () => {
   const [open, setOpen] = useState(false);
@@ -26,13 +22,10 @@ const Sort: React.FC = () => {
     open ? setOpen(false) : setOpen(true)
   };
 
+  //TODO прикинуть, как типизировать composedPaths
   useEffect(() => {
-    const handleClosePopup = (evt: MouseEvent) => {
-      const _event = evt as PopupClickType;
-
-      const path = _event.path || (_event.composedPath)
-
-      if (sortRef.current && !path.includes(sortRef.current)) {
+    const handleClosePopup = (evt: any) => {
+      if (sortRef.current && !evt.composedPath().includes(sortRef.current)) {
         setOpen(false);
       }
     };
