+++ src/Redux/Store.ts
@@ -14,4 +14,6 @@ const store = configureStore({
   },
 });
 
+export type RootState = ReturnType<typeof store.getState>
+
 export default store;
